Modify the implementation
Squashed commit of the following:

commit 339fe52e476504d6ff4ca6dae300fa26f0a09cae
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Wed Mar 11 15:32:41 2015 +0900

    Change the way to generate the test data

commit 9c11bc581bd1abd4571b39e5642651b26111d6f8
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Wed Mar 11 15:32:30 2015 +0900

    Modify the test

commit eb27c5aa3a4bb93d5a41e1c384a4b3571de83ead
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Wed Mar 11 13:55:24 2015 +0900

    Validate if the variance's norm is NaN or not

commit 83a8319e7fd29703f284fed65210748bede73c5d
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Tue Mar 10 15:43:35 2015 +0900

    Change variable type for a Map-class for a cluster tree

commit d4b67f1f06b82e29f7aa86161bda8bd2a823a342
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Thu Mar 5 18:09:40 2015 +0900

    Change the order of the arguments

commit f798742fe1200168749ffaf5ab248956b94e16fd
Author: Yu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Thu Mar 5 08:50:04 2015 +0000

    Modify the passes to the jar file

commit 032b1f051c35aff08fd81989438a5891a70df61c
Author: Yuu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Wed Mar 4 10:37:40 2015 +0900

    Add test for SparseVector

commit 8f9c9be2a05b8e92bd226b3432be0847080eb67e
Author: Yuu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Fri Feb 13 17:28:43 2015 +0900

    Change how to determine the maximum number of nodes

commit 60c7b3e63f215644734db030461c07584c0e31fc
Author: Yuu ISHIKAWA <yuu.ishikawa@gmail.com>
Date:   Fri Feb 13 17:28:16 2015 +0900

    Rename variables and change how to show the result

+++ src/test/scala/org/apache/spark/mllib/clustering/HierarchicalClusteringSuite.scala
@@ -96,7 +96,7 @@ class HierarchicalClusteringSuite extends FunSuite with LocalSparkContext {
     assert(clusters(1).center === Vectors.dense(49.5, 49.5))
     assert(clusters(1).records === 100)
 
-    val data2 = seed.map(v => ((v.apply(0) / 25).toInt + 1, v.toBreeze))
+    val data2 = seed.map(v => ((v.apply(0) / 25).toLong + 1L, v.toBreeze))
     val clusters2 = algo.summarizeAsClusters(data2)
     assert(clusters2.size === 4)
     assert(clusters2(1).center === Vectors.dense(12.0, 12.0))
@@ -112,8 +112,8 @@ class HierarchicalClusteringSuite extends FunSuite with LocalSparkContext {
   test("getChildrenCenter") {
     val algo = new HierarchicalClustering
     val centers = Map(
-      2 -> Vectors.dense(1.0, 1.0).toBreeze,
-      3 -> Vectors.dense(2.0, 2.0).toBreeze
+      2L -> Vectors.dense(1.0, 1.0).toBreeze,
+      3L -> Vectors.dense(2.0, 2.0).toBreeze
     )
     val initNextCenters = algo.initChildrenCenter(centers)
     assert(initNextCenters.size === 4)
@@ -122,7 +122,7 @@ class HierarchicalClusteringSuite extends FunSuite with LocalSparkContext {
 
   test("should divide clusters") {
     val algo = new HierarchicalClustering
-    val seed = (0 to 99).map(i => ((i / 50).toInt + 2, Vectors.dense(i, i).toBreeze))
+    val seed = (0 to 99).map(i => ((i / 50) + 2L, Vectors.dense(i, i).toBreeze))
     val data = sc.parallelize(seed)
     val clusters = algo.summarizeAsClusters(data)
     val newClusters = algo.getDividedClusters(data, clusters)
@@ -141,16 +141,16 @@ class HierarchicalClusteringSuite extends FunSuite with LocalSparkContext {
   test("should assign each data to new clusters") {
     val algo = new HierarchicalClustering
     val seed = Seq(
-      (2, Vectors.dense(0.0, 0.0)), (2, Vectors.dense(1.0, 1.0)), (2, Vectors.dense(2.0, 2.0)),
-      (2, Vectors.dense(3.0, 3.0)), (2, Vectors.dense(4.0, 4.0)), (2, Vectors.dense(5.0, 5.0)),
-      (3, Vectors.dense(6.0, 6.0)), (3, Vectors.dense(7.0, 7.0)), (3, Vectors.dense(8.0, 8.0)),
-      (3, Vectors.dense(9.0, 9.0)), (3, Vectors.dense(10.0, 10.0)), (3, Vectors.dense(11.0, 11.0))
+      (2L, Vectors.dense(0.0, 0.0)), (2L, Vectors.dense(1.0, 1.0)), (2L, Vectors.dense(2.0, 2.0)),
+      (2L, Vectors.dense(3.0, 3.0)), (2L, Vectors.dense(4.0, 4.0)), (2L, Vectors.dense(5.0, 5.0)),
+      (3L, Vectors.dense(6.0, 6.0)), (3L, Vectors.dense(7.0, 7.0)), (3L, Vectors.dense(8.0, 8.0)),
+      (3L, Vectors.dense(9.0, 9.0)), (3L, Vectors.dense(10.0, 10.0)), (3L, Vectors.dense(11.0, 11.0))
     ).map { case (idx, vector) => (idx, vector.toBreeze)}
     val newClusters = Map(
-      4 -> new ClusterTree(Vectors.dense(1.0, 1.0), 3, Vectors.dense(1.0, 1.0)),
-      5 -> new ClusterTree(Vectors.dense(4.0, 4.0), 3, Vectors.dense(1.0, 1.0)),
-      6 -> new ClusterTree(Vectors.dense(7.0, 7.0), 3, Vectors.dense(1.0, 1.0)),
-      7 -> new ClusterTree(Vectors.dense(10.0, 10.0), 3, Vectors.dense(1.0, 1.0))
+      4L -> new ClusterTree(Vectors.dense(1.0, 1.0), 3, Vectors.dense(1.0, 1.0)),
+      5L -> new ClusterTree(Vectors.dense(4.0, 4.0), 3, Vectors.dense(1.0, 1.0)),
+      6L -> new ClusterTree(Vectors.dense(7.0, 7.0), 3, Vectors.dense(1.0, 1.0)),
+      7L -> new ClusterTree(Vectors.dense(10.0, 10.0), 3, Vectors.dense(1.0, 1.0))
     )
     val data = sc.parallelize(seed)
     val result = algo.updateClusterIndex(data, newClusters).collect().toSeq
